AJD.jl Documentation

Documentation for the AJD.jl Package to perform Approximate Joint Diagonalization (AJD) on multiple matrices.

AJD.ajd_benchmark

AJD.check_input

AJD.create_linear_filter

AJD.diagonalize

AJD.frobenius_offdiag_normation

AJD.get_diag_elements

AJD.get_offdiag_elements

AJD.get_y_fdiag

AJD.get_z_fdiag

AJD.is_commuting

AJD.is_same_size

AJD.isstrictly_diagonally_dominant

AJD.jdiag_cardoso

AJD.jdiag_edourdpineau

AJD.jdiag_gabrieldernbach!

AJD.random_commuting_matrices

AJD.random_matrices

AJD.random_normal_commuting_matrices

AJD.random_symmetric_matrices

AJD.ajd_benchmark — Method

ajd_benchmark(n_dims::Int, n_matrices::Int)

Run benchmark of implemented algorithms with random inputs. Prints basic overview of median execution times. Returns BenchmarkGroup containing detailed results.

AJD.check_input — Method

Check for valid input of diagonalize().

AJD.create_linear_filter — Method

Create LinearFilter object as introduced by Diagonalizations.jl. Output of AJD.jl follows convention of Diagonalizations.jl and produces a LinearFilter.

AJD.diagonalize — Method

diagonalize(
    A::Vector{<:AbstractMatrix{<:Number}};
    algorithm::String = "jdiag_gabrieldernbach",
    max_iter::Int = 1000,
    threshold::AbstractFloat = eps()
    )

Calculate joint diagonalization of multiple input matrices using the requested algorithms.

Main function of the AJD package. Implemented algorithms at this point in time are limited to the JDiag algorithm in different versions. Input of matrices to be diagonalized need to be a vector of matrices. The matrices can be of types Float64 or Complex. Limitations of the different implementations apply.

Supported algorithms are jdiag_gabrieldernbach, jdiag_cardoso and jdiag_edourdpineau. See the Getting Started Guide for information on the algorithms.

AJD.frobenius_offdiag_normation — Method

frobenius_off_diag_normation(A::Array{<:Number,3})

Input

A: Vector of matrices

Takes an array namely the Array of matrices Ak and gets the offdiagonal elements and applies the frobenius norm (∑ |a{i,j}|^{2}). Used for the jdiag_gabrieldernbach and FFD algorithm.

AJD.get_diag_elements — Method

get_diag_elements(A::Array)

Input

A: Vector of matrices

Takes an array of matrices, takes the diagonal elements and returns the diagonal elements as a diagonal matrix D.

Examples

```jldoctest julia> A = ones(3,3,3);

julia> AJD.getdiagelements(A) 3×3×3 Array{Float64, 3}: [:, :, 1] = 1.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 1.0

[:, :, 2] = 1.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 1.0

[:, :, 3] = 1.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 1.0```

AJD.get_offdiag_elements — Method

get_offdiag_elements(A::Array{<:Number,3})

Input

A: Vector of matrices

Takes an array of matrices, copies it and sets the diagonal to zero to get the matrix E with only the offdiagonal elements of A.

Examples

```jldoctest julia> A = ones(3,3,3);

julia> AJD.getoffdiagelements(A) 3×3×3 Array{Float64, 3}: [:, :, 1] = 0.0 1.0 1.0 1.0 0.0 1.0 1.0 1.0 0.0

[:, :, 2] = 0.0 1.0 1.0 1.0 0.0 1.0 1.0 1.0 0.0

[:, :, 3] = 0.0 1.0 1.0 1.0 0.0 1.0 1.0 1.0 0.0```

AJD.get_y_fdiag — Method

get_y_fdiag(D::AbstractArray{<:Number}, E::AbstractArray{<:Number}, i::Int, j::Int)

Input

D: Diagonal Matrix with offdiagonal elements set to zero

E: Diagonal Matrix with diagonal elements set to zero

i,j: Denotes the indexes of the matrices D and E

Calculates the factor y_ij which is defined by: math ∑_{k} D_{j,j}^{k}E_{j,i}^{k}

AJD.get_z_fdiag — Method

get_z_fdiag(D::AbstractArray{<:Number}, i::Int, j::Int)

Input

D: Diagonal Matrix with offdiagonal elements set to zero

i,j: Denotes the indexes the matrix D

Calculates the factor z_ij which is defined by: ∑_{k} D_{i,i}^{k}D_{j,j}^{k}

AJD.is_commuting — Method

is_commuting(A::AbstractMatrix, B::AbstractMatrix)

Input:

A: AbstractMatrix of dimension nxn

B: AbstractMatrix of dimension nxn

Check if two matrices A, B are commuting. A * B = B * A must hold.

AJD.is_same_size — Method

is_same_size(A::AbstractMatrix, B::AbstractMatrix)

Input:

A: AbstractMatrix of variable size

B: AbstractMatrix of variable size

Will return true if dimension of A and B is matching otherwise false.

AJD.isstrictly_diagonally_dominant — Method

isstrictly_diagonally_dominant(A::AbstractMatrix)

Input:

A: AbstractMatrix

Used for the FFDiag Algorithm to define whether the Matrix A is strictly diagonally dominant and therefore has an Inverse or not. A matrix is strictly dominant if:|a_{ii}| > \sum |a_{ij}|, i ≠ j

AJD.jdiag_cardoso — Method

jdiag_cardoso(M,jthresh)

Only works for matrix with real valued entries. Based on Matlab Code by Cardoso.

Input:

A is a mxnm matrix,(A1,...,An), each with dimension mxm

thresh is a threshold for approximation stop, normally = 10e-8.

Output:

V : is a mxm matrix, which accumulates givens rotations G in each iteration.

A : is a mxnm matrix, which contains [VA1V',...,VAnV']

iter: accumulates the iteration numbers

AJD.jdiag_edourdpineau — Method

jdiag_edourdpineau(X::Vector{M}; iter=100, eps=1e-3) 
    where {T<:Union{Real,Complex},M<:AbstractMatrix{T}}

Diagonalize a set of matrices using the Jacobi method ("Jacobi Angles for Simultaneous Diagonalization"). Code adapted from Edouardpineaus Python implementation

AJD.jdiag_gabrieldernbach! — Method

(1) jdiag_gabrieldernbach(A::Vector{Matrix{Float64}}; threshold = eps(), max_iter = 1000)

JDiag algorithm based on the implementation by Gabrieldernbach in Python.

Source: https://github.com/gabrieldernbach/approximatejointdiagonalization/blob/master/jade/jade_cpu.py

(2) jdiag_gabrieldernbach(A::Vector{Matrix{ComplexF64}}; threshold = eps(), max_iter = 1000)

JDiag algorithm for complex matrices based on the implementation by Gabrieldernbach in Python, the Cardoso Paper and the code of https://github.com/edouardpineau/Time-Series-ICA-with-SOBI-Jacobi.

AJD.random_commuting_matrices — Method

Generate m random commuting matrices of size n x n These will produce all real rotation matrices using the Jacobi method

Mi * Mj = Mj * Mi for all i,j

AJD.random_matrices — Method

random_matrices(n::Int, m::Int)

Generate m random matrices of size n x n

AJD.random_normal_commuting_matrices — Method

random_normal_commuting_matrices(n::Int, m::Int; complex::Bool=false)

Generate m random normal commuting matrices of size n x n These can be exactly diagonalized

Mi * Mj = Mj * Mi for all i,j Mi*Mi' = Mi'*Mi for all i

AJD.random_symmetric_matrices — Method

Generate m random symmetric matrices of size n x n