AJD

Documentation for AJD.

AJD.QDiag

AJD.ffdiag

AJD.plot_convergence

AJD.QDiag — Function

QDiag(C_0, C, weights, approach, tolerance, maximum_iteration, return_iteration_errors
    random_number_generator = Xoshiro()) -> AbstractArray{<:Real}

This is an implementation of the algorithm introduced in: Vollgraf, Roland & Obermayer, Klaus. (2006). Quadratic optimization for simultaneous matrix diagonalization. Signal Processing, IEEE Transactions on. 54. 3270 - 3278. 10.1109/TSP.2006.877673. Namings of the parameters are consistent with the paper for easier understanding.

Arguments

C_0::AbstractArray{<:Real}: This is the sphering matrix introduced as $C^{(0)}$.

C::AbstractArray{<:Real}: This is the set of matrices that is introduced as $C$

weights::AbstractArray{<:Real}:: This is the weights vector $mathbf{lpha}$. It will be later normalized such that $sum{k=1}{K} lpha_k = 1$.

approach::String: This is the flag for approaches "NK3" and "N5" introduced in the paper.

tolerance::Real: This is the tolerance for the error.

maximum_iteration::Integer: Maximum number of iterations.

returniterationerrors::Bool: If true error log wil be returned.

randomnumbergenerator::Xoshiro: This is random number generator for matrix $W$. Default generator generates random numbers based on default_rng() but a seed can be introduced by the user. This argument is added for testing purposes.

AJD.ffdiag — Function

ffdiag(C0, runs=100, tol=1e-9) -> AbstractArray{<:Real}, Matrix{<:Real}, Array{<:Real}

Compute the transformation matrix that diagonalizes a set of symmetric matrices

This is an Implementation of the Algorithm introduced in: Ziehe, Andreas; Laskov, Pavel; Nolte, Guido; Müller Klaus-Robert. (2004). A Fast Algorithm for Joint Diagonalization with Non-orthogonal Transformations and its Application to Blind Source Separation. Journal of Machine Learning Research 5 (2004) 777–800.

The function returns the diagonalized set of matrices, the diagonalization matrix and an array of diagonlization errors per iteration.

Arguments

C0::AbstractArray{<:Real}: This is a set of matrices to be diagonalized.

runs::Int: The maximum number of iterations. The default is max 100 iterations.

tol::Float64: The tolerance for the error. The default is 1e-9.

AJD.plot_convergence — Method

plot_convergence(errs, label)

Plot the convergence error over iterations on a logarithmic scale.

Arguments

errs::AbstractArray{<:Real} An array with the errors, where the index corresponds to the number of iteration.

label::String Label of the figure.