AJD.jl Documentation

Documentation for the AJD.jl Package to perform Approximate Joint Diagonalization (AJD) on multiple matrices.

AJD.ajd_benchmark

AJD.create_linear_filter

AJD.diagonalize

AJD.jdiag_cardoso

AJD.jdiag_edourdpineau

AJD.jdiag_gabrieldernbach!

AJD.random_commuting_matrices

AJD.random_matrices

AJD.random_normal_commuting_matrices

AJD.random_symmetric_matrices

AJD.ajd_benchmark — Method

ajd_benchmark(n_dims::Int, n_matrices::Int)

Run benchmark of implemented algorithms with random inputs. Prints basic overview of median execution times. Returns BenchmarkGroup containing detailed results.

AJD.create_linear_filter — Method

Create LinearFilter object as introduced by Diagonalizations.jl. Output of AJD.jl follows convention of Diagonalizations.jl and produces a LinearFilter.

AJD.diagonalize — Method

diagonalize(
    A::Vector{<:AbstractMatrix{<:Number}};
    algorithm::String = "jdiag_gabrieldernbach",
    max_iter::Int = 1000,
    threshold::AbstractFloat = eps()
    )

Calculate joint diagonalization of multiple input matrices using the requested algorithms.

Main function of the AJD package. Implemented algorithms at this point in time are limited to the JDiag algorithm in different versions. Input of matrices to be diagonalized need to be a vector of matrices. The matrices can be of types Float64 or Complex. Limitations of the different implementations apply.

Supported algorithms are jdiag_gabrieldernbach, jdiag_cardoso and jdiag_edourdpineau. See the Getting Started Guide for information on the algorithms.

AJD.jdiag_cardoso — Method

jdiag_cardoso(M,jthresh)

Only works for matrix with real valued entries. Based on Matlab Code by Cardoso.

Input:

A is a mxnm matrix,(A1,...,An), each with dimension mxm

thresh is a threshold for approximation stop, normally = 10e-8.

Output:

V : is a mxm matrix, which accumulates givens rotations G in each iteration.

A : is a mxnm matrix, which contains [VA1V',...,VAnV']

iter: accumulates the iteration numbers

AJD.jdiag_edourdpineau — Method

jdiag_edourdpineau(X::Vector{M}; iter=100, eps=1e-3) 
    where {T<:Union{Real,Complex},M<:AbstractMatrix{T}}

Diagonalize a set of matrices using the Jacobi method ("Jacobi Angles for Simultaneous Diagonalization"). Code adapted from Edouardpineaus Python implementation

AJD.jdiag_gabrieldernbach! — Method

(1) jdiag_gabrieldernbach(A::Vector{Matrix{Float64}}; threshold = eps(), max_iter = 1000)

JDiag algorithm based on the implementation by Gabrieldernbach in Python.

Source: https://github.com/gabrieldernbach/approximatejointdiagonalization/blob/master/jade/jade_cpu.py

(2) jdiag_gabrieldernbach(A::Vector{Matrix{ComplexF64}}; threshold = eps(), max_iter = 1000)

JDiag algorithm for complex matrices based on the implementation by Gabrieldernbach in Python, the Cardoso Paper and the code of https://github.com/edouardpineau/Time-Series-ICA-with-SOBI-Jacobi.

AJD.random_commuting_matrices — Method

Generate m random commuting matrices of size n x n These will produce all real rotation matrices using the Jacobi method

Mi * Mj = Mj * Mi for all i,j

AJD.random_matrices — Method

random_matrices(n::Int, m::Int)

Generate m random matrices of size n x n

AJD.random_normal_commuting_matrices — Method

random_normal_commuting_matrices(n::Int, m::Int; complex::Bool=false)

Generate m random normal commuting matrices of size n x n These can be exactly diagonalized

Mi * Mj = Mj * Mi for all i,j Mi*Mi' = Mi'*Mi for all i

AJD.random_symmetric_matrices — Method

Generate m random symmetric matrices of size n x n