AJD.jl Documentation

Documentation for the AJD.jl Julia Package to perform Approximate Joint Diagonalization (AJD) on multiple matrices.

Contents

Getting Started Guide

Detailed documentation

Theory

Test Data Generation

References

Main AJD.jl function

Below the main function of the AJD.jl package diagonlaize() is introduced. For the entire documentation, including non exported functions, of the package refer to the detailed docs page.

AJD.diagonalize — Method

(1)
    diagonalize(
        M::Vector{<:AbstractMatrix{<:Number}};
        algorithm::AbstractDiagonalization = JDiagEdourdPineau(),
        max_iter::Int = 1000,
        threshold::AbstractFloat = eps(),
    )

(2)
    diagonalize(
        M::Vector{<:AbstractMatrix{<:Number}},
        only_plot::Symbol;
        algorithm::AbstractDiagonalization = JDiagGabrielDernbach(),
        max_iter::Int = 1000,
        threshold::AbstractFloat = eps(),
    )

(3)
    diagonalize(
        benchmark::Symbol,
        n_dims::Int,
        n_matrices::Int,
    )

Calculate joint diagonalization of multiple input matrices $M_k$.

Main function of the AJD package. Implemented algorithms are JDiag and FFDiag. Input of matrices $M_k$ need to be a vector of matrices. The matrices can be of types Float64 or Complex.

See the Getting Started Guide for information on the algorithms.

Dispatch (1)

Inputs

M: Vector of matrices (requiered).

algorithm: Selected algorithm from JDiagGabrielDernbach(), JDiagEdourdPineau(), JDiagCardoso() or FFDiag().

max_iter: Maximum iteration step as integer.

threshold: Desired threshold minimizing the off-diagonal elements.

Output

Return LinearFilter object. Filter fil contains filter matrix fil.F and the inverse fil.iF.

Dispatch (2)

Inputs

Additional symbol used to generate overview plot of the result. Use diagonalize(M, :plot).

Output

Overview plot. Shows heatmap of the filter matrix, heatmap of the mean of the diagonalized matrices and the vonvergence behaviour of the algorithm.

Dispatch (3)

Inputs

Symbol :benchmark used as diagonalize(:benchmark, 10, 10).

Automatic benchmark is run comparing JDiag and FFDiag algorithms. Using n_dims $\times$ n_dims matrices of count n_matrices.

Output

BenchmarkGroup of package BenchmarkTools comparing the algorithms using diffrent types of test data.