AJD.jl Documentation

Documentation for the AJD.jl Package to perform Approximate Joint Diagonalization (AJD) on multiple matrices.

AJD.diagonalize

AJD.jdiag_cardoso

AJD.jdiag_edourdpineau

AJD.jdiag_gabrieldernbach

AJD.off_diag_normation

AJD.diagonalize — Method

diagonalize(
    A::Vector{<:AbstractMatrix{<:Union{Float64, ComplexF64}}};
    algorithm::String)

Diagonalize input matrices using the requested algorithm.

Main function of the AJD package. Implemented algorithms at this point in time are limited to the JDiag algorithm in different versions. Input of matrices to be diagonalized need to be a vector of matrices. The matrices can be of types Float64 or Complex. Limitations of the different implementations apply.

Supported algorithms are jdiag_gabrieldernbach, jdiag_cardoso and jdiag_edourdpineau. See the Getting Started Guide for information on the algorithms.

AJD.jdiag_cardoso — Method

jdiag_cardoso(M,jthresh)

Only works for matrix with real valued entries. Based on Matlab Code by Cardoso.

Input:

A is a mxnm matrix,(A1,...,An), each with dimension mxm

thresh is a threshold for approximation stop, normally = 10e-8.

Output:

V : is a mxm matrix, which accumulates givens rotations G in each iteration.

A : is a mxnm matrix, which contains [VA1V',...,VAnV']

iter: accumulates the iteration numbers

AJD.jdiag_edourdpineau — Method

jdiag_edourdpineau(X::Vector{M}; iter=100, eps=1e-3) where {T<:Union{Real,Complex},M<:AbstractMatrix{T}}

Diagonalize a set of matrices using the Jacobi method.
"Jacobi Angles for Simultaneous Diagonalization"
Code adapted from:
https://github.com/edouardpineau/Time-Series-ICA-with-SOBI-Jacobi

AJD.jdiag_gabrieldernbach — Method

(1) jdiag_gabrieldernbach(A::Vector{Matrix{Float64}}; threshold = eps(), max_iter = 1000)

JDiag algorithm based on the implementation by Gabrieldernbach in Python.

Source: https://github.com/gabrieldernbach/approximatejointdiagonalization/blob/master/jade/jade_cpu.py

(2) jdiag_gabrieldernbach(A::Vector{Matrix{ComplexF64}}; threshold = eps(), max_iter = 1000)

JDiag algorithm for complex matrices based on the implementation by Gabrieldernbach in Python, the Cardoso Paper and the code of https://github.com/edouardpineau/Time-Series-ICA-with-SOBI-Jacobi.

AJD.off_diag_normation — Method

off_diag_normation(A::Array)
**input**
A: Vector of communting matrices with index k

Takes an array namely the Array of matrices Ak and gets the offdiagonal elements, norms them and adds them up to return normation. Used for the `jdiaggabrieldernbach` algorithm