AJD.jl Documentation

Documentation for the AJD.jl Julia Package to perform Approximate Joint Diagonalization (AJD) on multiple matrices.

Contents

Getting Started Guide

Detailed documentation

Theory

References

Main AJD.jl functions

Below the main function of the AJD.jl package diagonlaize() is introduced below. For the entire documentation, including non exported functions, of the package refer to the detailed docs page.

AJD.diagonalize — Method

diagonalize(
    A::Vector{<:AbstractMatrix{<:Number}};
    algorithm::String = "jdiag_gabrieldernbach",
    max_iter::Int = 1000,
    threshold::AbstractFloat = eps(),
    plot_matrix::Bool = false,
    plot_convergence::Bool = false
    )

Calculate joint diagonalization of multiple input matrices $M_k$.

Main function of the AJD package. Implemented algorithms are JDiag and FFDiag. Input of matrices $M_k$ need to be a vector of matrices. The matrices can be of types Float64 or Complex.

Supported algorithms are jdiag_gabrieldernbach, jdiag_cardoso, jdiag_edourdpineau and ffdiag. See the Getting Started Guide for information on the algorithms.