Detailed Docs

AJD.ALL_ALGORITHMS

AJD.COMLPLEX_ALGORITHMS

AJD.ajd_benchmark

AJD.check_input

AJD.create_linear_filter

AJD.diagonalize

AJD.diagonalize

AJD.frobenius_offdiag_norm

AJD.generate_correlation_matrix

AJD.generate_testdata

AJD.generate_testdata

AJD.get_diag_elements

AJD.get_offdiag_elements

AJD.get_y_fdiag

AJD.get_z_fdiag

AJD.is_commuting

AJD.is_same_size

AJD.isstrictly_diagonally_dominant

AJD.jdiag_cardoso

AJD.jdiag_edourdpineau

AJD.jdiag_gabrieldernbach!

AJD.plot_convergence_lineplot

AJD.plot_matrix_heatmap

AJD.random_commuting_matrices

AJD.random_matrices

AJD.random_normal_commuting_matrices

AJD.random_symmetric_matrices

AJD.ALL_ALGORITHMS — Constant

List of all algorithms keywords. Used in tests.

AJD.COMLPLEX_ALGORITHMS — Constant

List of all algorithms supporting complex matrices.

AJD.ajd_benchmark — Method

ajd_benchmark(n_dims::Int, n_matrices::Int)

Run benchmark of implemented algorithms with random inputs. Prints basic overview of median execution times. Returns BenchmarkGroup containing detailed results.

AJD.check_input — Method

Check for valid input of diagonalize().

AJD.create_linear_filter — Method

Create LinearFilter object as introduced by Diagonalizations.jl. Output of AJD.jl follows convention of Diagonalizations.jl and produces a LinearFilter.

AJD.diagonalize — Method

diagonalize(
    A::Vector{<:AbstractMatrix{<:Number}};
    algorithm::String = "jdiag_gabrieldernbach",
    max_iter::Int = 1000,
    threshold::AbstractFloat = eps(),
    plot_matrix::Bool = false,
    plot_convergence::Bool = false
    )

Calculate joint diagonalization of multiple input matrices $M_k$.

Main function of the AJD package. Implemented algorithms are JDiag and FFDiag. Input of matrices $M_k$ need to be a vector of matrices. The matrices can be of types Float64 or Complex.

Supported algorithms are jdiag_gabrieldernbach, jdiag_cardoso and jdiag_edourdpineau. See the Getting Started Guide for information on the algorithms. Test

$M_k$.

AJD.frobenius_offdiag_norm — Method

frobenius_offdiag_norm(A::AbstractArray{T,3})::Real where {T<:Number}

Input

A: Vector of matrices

Takes an array namely the Array of matrices Ak and gets the offdiagonal elements and applies the frobenius norm (``\sum |a{i,j}|^{2}``).

AJD.generate_correlation_matrix — Method

generate_correlation_matrix(signal_one_data,signal_two_data)

signalonedata: Array of dimension (NxM)

signaltwodata: Array of dimension (NxM)

Calculates correlation matrix between observations $x_i(t)$ and $x_i(t+τ)$.

AJD.generate_testdata — Method

generate_testdata(signal_sources::AbstractArray{<:Function}, mixing_matrix::AbstractMatrix{<:Number}; <keyword_arguments>)

signal_sources: Array of anonymous functions for generating time series data of the uncorrelated signals s of BSS e.g. [ s1 = x-> 1.4*sin(2x), s2 = 2.2sin(x)]

mixing matrix: mixing matrix by which the signals $s_j$ are multiplied to get the measurements/observations $x_i$

First Calculates from a given array of functions resolved in the time domain, which simulate the uncorrelated signals $s_j$ and a mixing matrix A, the measurements $x_i$ with:

x_i(t) = \sum_{t = 1}^{T} a_{i,j} * s_j(t).

Then time delayed correlation matrix between observations mentioned in [source] for a specified number in no_of_corr.

Arguments

delay::Number = 1: time delay between signals

sample_time::Number = 10: length of single time series (same for all observations)

noofsamples::Int = 100: number of observations in the time series

noofcor::Int = 10: number of observations made over the entire measurement

For numbers to generate test data set of timeresolved function look here: https://doi.org/10.21595/jve.2021.21961 p.1709

AJD.generate_testdata — Method

generate_testdata(signal_sources::AbstractArray; delay::Number = 10, no_of_segments::Int = 10)

signalsources: Matrix of rowwise signals [``x1$;$x2$;...;$xn``]

delay: Time/index shift between observations to be correlated

noof:segments: Number of correlation matrices - 1 to be calculated. Puts signalsources into even segments to be correlated. If the number leads to uneven correlation will throw an error.

Generate Correlation Matrices for discrete observations $x_i$.

Known Issue

If your data has a lot of zeros inside the observations setting noofsegements too high will lead to NaN values since the variance of a vector of zeros is zero! You might want to manipulate your data or change the number of segments to be less!

AJD.get_diag_elements — Method

get_diag_elements(A::Array)

Input

A: Vector of matrices

Takes an array of matrices, takes the diagonal elements and returns the diagonal elements as a diagonal matrix D.

Examples

```jldoctest julia> A = ones(3,3,3);

julia> AJD.getdiagelements(A) 3×3×3 Array{Float64, 3}: [:, :, 1] = 1.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 1.0

[:, :, 2] = 1.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 1.0

[:, :, 3] = 1.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 1.0```

AJD.get_offdiag_elements — Method

get_offdiag_elements(A::Array{<:Number,3})

Input

A: Vector of matrices

Takes an array of matrices, copies it and sets the diagonal to zero to get the matrix E with only the offdiagonal elements of A.

Examples

```jldoctest julia> A = ones(3,3,3);

julia> AJD.getoffdiagelements(A) 3×3×3 Array{Float64, 3}: [:, :, 1] = 0.0 1.0 1.0 1.0 0.0 1.0 1.0 1.0 0.0

[:, :, 2] = 0.0 1.0 1.0 1.0 0.0 1.0 1.0 1.0 0.0

[:, :, 3] = 0.0 1.0 1.0 1.0 0.0 1.0 1.0 1.0 0.0```

AJD.get_y_fdiag — Method

get_y_fdiag(D::AbstractArray{<:Number}, E::AbstractArray{<:Number}, i::Int, j::Int)

Input

D: Diagonal Matrix with offdiagonal elements set to zero

E: Diagonal Matrix with diagonal elements set to zero

i,j: Denotes the indexes of the matrices D and E

Calculates the factor $y_ij$ which is defined by: $∑_{k} D_{j,j}^{k}E_{j,i}^{k}$

AJD.get_z_fdiag — Method

get_z_fdiag(D::AbstractArray{<:Number}, i::Int, j::Int)

Input

D: Diagonal Matrix with offdiagonal elements set to zero

i,j: Denotes the indexes the matrix D

Calculates the factor $z_ij$ which is defined by: $∑_{k} D_{i,i}^{k}D_{j,j}^{k}$

AJD.is_commuting — Method

is_commuting(A::AbstractMatrix, B::AbstractMatrix)

Input:

A: AbstractMatrix of dimension nxn

B: AbstractMatrix of dimension nxn

Check if two matrices A, B are commuting. A * B = B * A must hold.

AJD.is_same_size — Method

is_same_size(A::AbstractMatrix, B::AbstractMatrix)

Input:

A: AbstractMatrix of variable size

B: AbstractMatrix of variable size

Will return true if dimension of A and B is matching otherwise false.

AJD.isstrictly_diagonally_dominant — Method

isstrictly_diagonally_dominant(A::AbstractMatrix)

Input:

A: AbstractMatrix

Used for the FFDiag Algorithm to define whether the Matrix A is strictly diagonally dominant and therefore has an Inverse or not. A matrix is strictly dominant if:$|a_{ii}| > \sum |a_{ij}|, i ≠ j$

AJD.jdiag_cardoso — Method

jdiag_cardoso(M,jthresh)

Only works for matrix with real valued entries. Based on Matlab Code by Cardoso.

Input:

A is a mxnm matrix,(A1,...,An), each with dimension mxm

thresh is a threshold for approximation stop, normally = 10e-8.

Output:

V : is a mxm matrix, which accumulates givens rotations G in each iteration.

A : is a mxnm matrix, which contains [VA1V',...,VAnV']

iter: accumulates the iteration numbers

AJD.jdiag_edourdpineau — Method

jdiag_edourdpineau(X::Vector{M}; iter=100, eps=1e-3)
    where {T<:Union{Real,Complex},M<:AbstractMatrix{T}}

Diagonalize a set of matrices using the Jacobi method ("Jacobi Angles for Simultaneous Diagonalization"). Code adapted from Edouardpineaus Python implementation

AJD.jdiag_gabrieldernbach! — Method

(1) jdiag_gabrieldernbach(A::Vector{Matrix{Float64}}; threshold = eps(), max_iter = 1000)

JDiag algorithm based on the implementation by Gabrieldernbach in Python.

Source: https://github.com/gabrieldernbach/approximatejointdiagonalization/blob/master/jade/jade_cpu.py

(2) jdiag_gabrieldernbach(A::Vector{Matrix{ComplexF64}}; threshold = eps(), max_iter = 1000)

JDiag algorithm for complex matrices based on the implementation by Gabrieldernbach in Python, the Cardoso Paper and the code of https://github.com/edouardpineau/Time-Series-ICA-with-SOBI-Jacobi.

AJD.plot_convergence_lineplot — Method

plot_convergence_lineplot(error_array::AbstractArray, name::String)

Plot the convergence error as recorded during the algorithm execution.

AJD.plot_matrix_heatmap — Method

plot_matrix_heatmap(filter::AbstractMatrix, diag_matrices::AbstractMatrix)

Plot a heatmap of the calculated filter matrix and the mean of the diagonlized matrices. Complex matrices are reduced to real matrices for the plot.

AJD.random_commuting_matrices — Method

Generate m random commuting matrices of size n x n These will produce all real rotation matrices using the Jacobi method

Mi * Mj = Mj * Mi for all i,j

AJD.random_matrices — Method

random_matrices(n::Int, m::Int)

Generate m random matrices of size n x n

AJD.random_normal_commuting_matrices — Method

random_normal_commuting_matrices(n::Int, m::Int; complex::Bool=false)

Generate m random normal commuting matrices of size n x n These can be exactly diagonalized

Mi * Mj = Mj * Mi for all i,j Mi*Mi' = Mi'*Mi for all i

AJD.random_symmetric_matrices — Method

Generate m random symmetric matrices of size n x n