Detailed Docs

AJD.ALL_ALGORITHMS

AJD.COMLPLEX_ALGORITHMS

AJD.ajd_benchmark

AJD.check_input

AJD.create_linear_filter

AJD.diagonalize

AJD.diagonalize

AJD.frobenius_offdiag_norm

AJD.generate_correlation_matrix

AJD.generate_random_signals

AJD.generate_testdata

AJD.generate_testdata

AJD.get_diag_elements

AJD.get_diagonalization

AJD.get_offdiag_elements

AJD.get_plot

AJD.get_y_fdiag

AJD.get_z_fdiag

AJD.is_commuting

AJD.is_same_size

AJD.isstrictly_diagonally_dominant

AJD.jdiag_cardoso

AJD.jdiag_edourdpineau

AJD.jdiag_gabrieldernbach!

AJD.random_normal_commuting_matrices

AJD.ALL_ALGORITHMS — Constant

List of all algorithms keywords. Used in tests.

AJD.COMLPLEX_ALGORITHMS — Constant

List of all algorithms supporting complex matrices.

AJD.ajd_benchmark — Method

ajd_benchmark(n_dims::Int, n_matrices::Int)

Run benchmark of implemented algorithms with random inputs. Prints basic overview of median execution times. Returns BenchmarkGroup containing detailed results.

AJD.check_input — Method

Check for valid input of diagonalize().

AJD.create_linear_filter — Method

Create LinearFilter object as introduced by Diagonalizations.jl. Output of AJD.jl follows convention of Diagonalizations.jl and produces a LinearFilter.

AJD.diagonalize — Method

diagonalize(
    A::Vector{<:AbstractMatrix{<:Number}};
    algorithm::String = "jdiag_gabrieldernbach",
    max_iter::Int = 1000,
    threshold::AbstractFloat = eps(),
    plot_matrix::Bool = false,
    plot_convergence::Bool = false
    )

Calculate joint diagonalization of multiple input matrices $M_k$.

Main function of the AJD package. Implemented algorithms are JDiag and FFDiag. Input of matrices $M_k$ need to be a vector of matrices. The matrices can be of types Float64 or Complex.

Supported algorithms are jdiag_gabrieldernbach, jdiag_cardoso and jdiag_edourdpineau. See the Getting Started Guide for information on the algorithms.

AJD.frobenius_offdiag_norm — Method

frobenius_offdiag_norm(A::AbstractArray{T,3})::Real where {T<:Number}

A: Vector of matrices with size $n × n × k$

Takes the offdiagonal elements of an Array of matrices $A^k$ and applies the frobenius norm ($\sum |a_{i,j}|^{2}$).

AJD.generate_correlation_matrix — Method

generate_correlation_matrix(signal_one_data::AbstractArray,signal_two_data::AbstractArray)

signal_one_data: Array of dimension $n × m$

signal_two_data: Array of dimension $n × m$

Calculates correlation matrix between observations $x_i(t)$ and $x_i(t+τ)$.

AJD.generate_random_signals — Method

generate_random_signals(no_of_signals::Int, 
no_of_samples::Int; seed = Xoshiro(), signal_type = Float64)

*no_of_signals: used to declare how many signals $s_{j}$ are inside of the testdata *no_of_samples: used to declare how many samples each signal $s_j$ has *seed: seed for the rand function, might need to import the module Random *signal_type: default is Float64 but ComplexF64 is possible as well

Similar to random_commuting_matrices function. Gives the opportunity to generate testdata randomly and pluck it into the discrete version of generate_testdata.

AJD.generate_testdata — Method

generate_testdata(signal_sources::AbstractArray{<:Function}, 
mixing_matrix::AbstractMatrix{<:Number}; <keyword_arguments>)

signal_sources: Array of anonymous functions for generating time series data of the uncorrelated signals $s_j$ of BSS e.g. [ s1 = x-> 1.4*sin(2x), s2 = 2.2sin(x)]

mixing_matrix: mixing matrix by which the signals $s_j$ are multiplied to get the measurements/observations $x_i$

First Calculates from a given array of functions, which simulate the uncorrelated signals $s_j(t)$ and a mixing matrix A, the observations $x_i$:

$x_i(t) = \sum_{t = 1}^{T} a_{i,j} s_j(t)$.

Then a number of time delayed correlation matrices specified by no_of_corr is calculated.

Arguments

delay::Number = 1: time delay between signals

sample_time::Number = 10: length of single time series (same for all observations)

no_of_samples::Int = 100: number of observations made during sample_time

no_of_cor::Int = 10: number of observations made over the entire measurement

Example for signal_sources and mixing_matrix

signal_sources = [x->1.6sin(2pi*5x+5)+2sin(2pi*20x+27)+0.5sin(2pi*100x)+1,x->1.2(2pi*11x)+sin(2pi*2x)+0.7sin(2pi*111x+10)]
mixing_matrix = [0.32 -0.43; -1.31 0.34]

julia

AJD.generate_testdata — Method

generate_testdata(signal_sources::AbstractArray; 
delay::Number = 10, no_of_segments::Int = 10)

signal_sources: Matrix of rowwise signals [$x_1$; $x_2$;...; $x_n$]

delay: Time/index shift between observations to be correlated

no_of_segments: Puts signal_sources into even segments to be correlated. If the number leads to uneven correlation will throw a warning if show_warning is true

show_warning: If true will show a warning in case segments are uneven. Will lead to one less correlation matrix

Generate Correlation Matrices for discrete observations $x_i$.

Known Issue

If your data has a segment with variance close to 0 (e.g. due to all of the values being the same) the correlation matrix will have NaN values inside. Setting the number of segments to a lower value might help.

AJD.get_diag_elements — Method

get_diag_elements(A::Array)

A: Vector of matrices

Takes an array of matrices and returns the diagonal elements as a diagonal matrix D.

Examples

```julia julia> A = ones(3,3,3);

julia> AJD.getdiagelements(A) 3×3×3 Array{Float64, 3}: [:, :, 1] = 1.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 1.0

[:, :, 2] = 1.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 1.0

[:, :, 3] = 1.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 1.0```

AJD.get_diagonalization — Method

function get_diagonalization(
    A::Vector{<:AbstractMatrix{<:Number}};
    algorithm::String = "jdiag_gabrieldernbach",
    max_iter::Int = 1000,
    threshold::AbstractFloat = eps(),
    only_plot::Symbol = :no_plot
    )

Get the actual diagonalization. Function is seperated from diagonalize() to facilitate plotting functionality in the REPL and Pluto. All implemented algorithms are called from this function. To generate the error histories of the algorithm runs, as used for the plots, select only_plot=:plot. Input is checked here as well.

AJD.get_offdiag_elements — Method

get_offdiag_elements(A::Array{<:Number,3})

A: Vector of matrices

Takes an array of matrices and returns the offdiagonal elements of A.

Examples

```julia julia> A = ones(3,3,3);

julia> AJD.getoffdiagelements(A) 3×3×3 Array{Float64, 3}: [:, :, 1] = 0.0 1.0 1.0 1.0 0.0 1.0 1.0 1.0 0.0

[:, :, 2] = 0.0 1.0 1.0 1.0 0.0 1.0 1.0 1.0 0.0

[:, :, 3] = 0.0 1.0 1.0 1.0 0.0 1.0 1.0 1.0 0.0```

AJD.get_plot — Method

get_plot(
    filter::AbstractArray,
    diag_matrices::AbstractArray, 
    error_array::AbstractArray, 
    name::String)

In case of plot is user selected this generates heatmap plot of the filter matrix and the mean of diagonlaised matrices. Also a lineplot of the error history of the algorithm calculation is created. A combined plot is returned.

AJD.get_y_fdiag — Method

get_y_fdiag(D::AbstractArray{<:Number}, E::AbstractArray{<:Number}, i::Int, j::Int)

D: Diagonal Matrix with offdiagonal elements set to zero

E: Diagonal Matrix with diagonal elements set to zero

i,j: Denotes the indexes of the matrices D and E

Calculates the factor $y_{ij}$ which is defined by: $∑_{k} D_{j,j}^{k}E_{j,i}^{k}$

AJD.get_z_fdiag — Method

get_z_fdiag(D::AbstractArray{<:Number}, i::Int, j::Int)

D: Diagonal Matrix with offdiagonal elements set to zero

i,j: Denotes the indexes of matrix D

Calculates the factor $z_{ij}$ which is defined by: $∑_{k} D_{i,i}^{k}D_{j,j}^{k}$

AJD.is_commuting — Method

is_commuting(A::AbstractMatrix, B::AbstractMatrix)

A: AbstractMatrix of dimension $n × n$

B: AbstractMatrix of dimension $n × n$

Check if two matrices A, B are commuting. $AB = BA$ must hold.

AJD.is_same_size — Method

is_same_size(A::AbstractMatrix, B::AbstractMatrix)

A: AbstractMatrix of variable size

B: AbstractMatrix of variable size

Will return true if dimension of A and B is matching otherwise false.

AJD.isstrictly_diagonally_dominant — Method

isstrictly_diagonally_dominant(A::AbstractMatrix)

A: AbstractMatrix

Used for the FFDiag Algorithm to define whether the Matrix A is strictly diagonally dominant and therefore has an Inverse or not. A matrix is strictly dominant if: $|a_{ii}| > \sum |a_{ij}|, i ≠ j$

AJD.jdiag_cardoso — Method

jdiag_cardoso(M,jthresh)

Only works for matrix with real valued entries. Based on Matlab Code by Cardoso.

Input:

A is a $m × m × n$ matrix,($A_1,...,A_n$), each with dimension $m × m$

thresh is a threshold for approximation stop, normally = 10e-8.

Output:

V : is a $m × m$ matrix, which accumulates givens rotations G in each iteration.

A : is a $m × m × n$ matrix, which contains [$VA_1V'$,...,$VA_nV'$]

iter: accumulates the iteration numbers

AJD.jdiag_edourdpineau — Method

jdiag_edourdpineau(X::Vector{M}; iter=100, eps=1e-3)
    where {T<:Union{Real,Complex},M<:AbstractMatrix{T}}

Diagonalize a set of matrices using the Jacobi method ("Jacobi Angles for Simultaneous Diagonalization"). Code adapted from Edouardpineaus Python implementation

AJD.jdiag_gabrieldernbach! — Method

(1) jdiag_gabrieldernbach(A::Vector{Matrix{Float64}}; 
    threshold = eps(), max_iter = 1000)

JDiag algorithm based on the implementation by Gabrieldernbach in Python.

Source: Algorithm

(2) jdiag_gabrieldernbach(A::Vector{Matrix{ComplexF64}}; 
    threshold = eps(), max_iter = 1000)

JDiag algorithm for complex matrices based on the implementation by Gabrieldernbach in Python, the Cardoso Paper and the code of Algorithm

AJD.random_normal_commuting_matrices — Method

random_normal_commuting_matrices(n::Int, m::Int; complex::Bool=false)

Generate m random normal commuting matrices of size $n × n$ These can be exactly diagonalized

$M_i M_j = M_j M_i$ for all i,j $M_i M_i' = M_i' M_i$ for all i